autoca/app/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
django-celery-results==2.5.1
django-celery-beat==2.5.0
django-filter==23.5
django-crispy-forms==2.1
django-allauth==0.57.0
django-storages==1.14.2
django-ratelimit==3.0.1

# Database & Storage
psycopg2-binary==2.9.9  # Use binary for development
Pillow==10.1.0  # Image processing
boto3==1.34.0  # AWS S3 storage
openpyxl==3.1.2  # Excel export

# Celery & Redis
celery[redis]==5.3.6
redis==5.0.1

# ML & Data Processing
easyocr==1.7.1  # OCR
pytesseract==0.3.10  # Tesseract wrapper
pdf2image==1.16.3  # PDF to image conversion
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2  # IsolationForest
scipy==1.11.3

# Payment & Security
stripe==7.0.0
cryptography==41.0.7

# HTTP & APIs
requests==2.31.0
gunicorn==21.2.0  # Production server
python-multipart==0.0.6  # File uploads